refactored scroll top popup
