+++ src/components/atoms/index.ts
@@ -3,3 +3,4 @@ export * from './progressBar';
 export * from './assetIcon';
 export * from './writePostButton';
 export * from './newPostsPopup';
+export * from './scrollTopPopup';
