+++ src/components/atoms/scrollTopPopup/index.ts
@@ -1,0 +1,1 @@
+export * from './scrollTopPopup';
