+++ src/components/tabbedPosts/styles/tabbedPosts.styles.tsx
@@ -1,3 +1,4 @@
+import { ViewStyle } from 'react-native';
 import EStyleSheet from 'react-native-extended-stylesheet';
 
 export default EStyleSheet.create({
@@ -40,49 +41,10 @@ export default EStyleSheet.create({
     marginVertical: 16,
     color: '$primaryBlack',
   },
-  popupContainer: {
-    position: 'absolute',
-    top: 24,
-    left: 0,
-    right: 0,
-    alignItems: 'center',
-  },
-  popupContentContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: '$primaryBlue',
-    paddingHorizontal: 2,
-    paddingVertical: 2,
-    borderRadius: 32,
-  },
-  postedText: {
-    fontWeight: '500',
-    color: '$primaryGray',
-    marginLeft: 6,
-  },
-  scrollTopText: {
-    fontWeight: '500',
-    color: '$primaryGray',
-    marginHorizontal: 4,
-    marginBottom: 2,
-  },
   closeIcon: {
     color: '$primaryGray',
     margin: 0,
     padding: 6,
-  },
-  arrowUpIcon: {
-    color: '$primaryGray',
-    margin: 0,
-    marginHorizontal: 4,
-  },
-  popupImage: {
-    height: 24,
-    width: 24,
-    borderRadius: 12,
-    borderWidth: 2,
-    marginLeft: -8,
-    borderColor: '$primaryBlue',
   },
   moreIcon: {
     marginLeft: -4,
