+++ src/screens/waves/screen/wavesScreen.tsx
@@ -16,8 +16,7 @@ import { wavesQueries } from '../../../providers/queries';
 import { useAppSelector } from '../../../hooks';
 import WavesHeader from '../children/wavesHeader';
 import { PostTypes } from '../../../constants/postTypes';
-import ScrollTopPopup from '../../../components/tabbedPosts/view/scrollTopPopup';
-import { NewPostsPopup } from '../../../components/atoms';
+import { NewPostsPopup, ScrollTopPopup } from '../../../components/atoms';
 
 const SCROLL_POPUP_THRESHOLD = 5000;
 
@@ -173,6 +172,10 @@ const WavesScreen = ({ route }) => {
           onClose={() => {
             setPopupAvatars([]);
           }}
+        />
+        <ScrollTopPopup
+          enable={enableScrollTop}
+          onPress={_scrollTop}
         />
       </View>
       <PostOptionsModal ref={postOptionsModalRef} isVisibleTranslateModal={true} isWave={true} />
